% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/bemkl_supervised_multioutput_regression_variational_test.R
\name{bemkl_supervised_multioutput_regression_variational_test}
\alias{bemkl_supervised_multioutput_regression_variational_test}
\title{BEMKL algorithm uses the training model parameters on the test set.}
\usage{
bemkl_supervised_multioutput_regression_variational_test(Km, state)
}
\arguments{
\item{Km}{numeric matrix with data}

\item{state}{numeric matrix with data}
}
\value{
numeric vector with the predictions
}
\description{
\code{bemkl_supervised_multioutput_regression_variational_test} implements bemkl on the test data
}
