% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_immune_response.R
\name{compare_immune_response}
\alias{compare_immune_response}
\title{Comparison of the actual predictions based on different metrics with real patient response.}
\usage{
compare_immune_response(
  predictions_immune_response = NULL,
  real_patient_response,
  RNA.tpm,
  output_file_path,
  list_gold_standards,
  cancertype
)
}
\arguments{
\item{predictions_immune_response}{list that contains the predictions for each model}

\item{real_patient_response}{vector with two factors (NR,R)}

\item{RNA.tpm}{numeric matrix with data}

\item{output_file_path}{string with a file name}

\item{list_gold_standards}{string with gold standards names}

\item{cancertype}{specify cancer type}
}
\value{
ROC curves plots and barplots showing AUC values.
}
\description{
\code{compare_response} plots ROC curves and barplots showing the accuracy of the predictions
on the real patient response data. It receives as input the predicted immune response as well
as the real patient response (they should be provided with the same order of samples). Gold
standards are also plotted to consider them as reference. To compute these gold standards,
transcriptomics data in different formats is provided. An output file path should be given
to save the plots as pdfs. An additional input must be a list with the gold standards the
user wish to compare the prediction results.
}
