% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Ayers_expIS.R
\name{compute.Ayers_expIS}
\alias{compute.Ayers_expIS}
\title{Compute Expanded Immune signature}
\usage{
compute.Ayers_expIS(RNA.tpm)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=Expanded Immune signature score
}
\description{
\code{compute_ayersEI} computes Expanded Immune signature score as the arithmetic mean of genes included
in the Expanded Immune signature (Ayers et al., JCI, 2017)
}
