% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.chemokines.R
\name{compute.chemokines}
\alias{compute.chemokines}
\title{Compute chemokine score}
\usage{
compute.chemokines(RNA.tpm)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=chemokine score
}
\description{
\code{compute_chemokine} computes chemoine score as the PC1 score that results from applying PCA
to z-score expression of 12 chemokine genes (Messina et al., 2012).
}
