% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Davoli_IS.R
\name{compute.Davoli_IS}
\alias{compute.Davoli_IS}
\title{Compute Davoli immune signature}
\usage{
compute.Davoli_IS(RNA.tpm)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=Davoli immune signature
}
\description{
\code{compute_davoliIS} computes Davoli immune signature as the arithmetic mean of cytotoxic
immune infiltrate signature genes, after rank normalization (Davoli et al., 2017).
}
