% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.IFNy.R
\name{compute.IFNy}
\alias{compute.IFNy}
\title{Compute IFNy signature score}
\usage{
compute.IFNy(RNA.tpm)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=IFNy signature score
}
\description{
\code{compute_ayersIFNy} computes IFNy signature score as the arithmetic mean of genes included
in the IFN-γ signature (Ayers et al., JCI, 2017)
}
