% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.IMPRES.R
\name{compute.IMPRES}
\alias{compute.IMPRES}
\title{Compute Immuno-Predictive Score (IMPRES)}
\usage{
compute.IMPRES(RNA.tpm)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=IMPRES score
}
\description{
\code{compute_IMPRES} computes IMPRES score by applying logical comparison of checkpoint gene pairs
(Auslander et al., 2018).
}
