% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.MSI.R
\name{compute.MSI}
\alias{compute.MSI}
\title{Compute MSI score}
\usage{
compute.MSI(RNA.tpm)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=MSI score
}
\description{
\code{compute_MSI} computes MSI score by applying logical comparison of MSI-related gene pairs
(Fu et al., 2019).
}
