% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Roh_IS.R
\name{compute.Roh_IS}
\alias{compute.Roh_IS}
\title{Compute Roh immune score}
\usage{
compute.Roh_IS(RNA.tpm)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=Roh immune score
}
\description{
\code{compute_rohIS} computes Roh immune score as the geometric-mean of immune score genes
(Roh et al., 2017).
}
