% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Tcell_inflamed.R
\name{compute.Tcell_inflamed}
\alias{compute.Tcell_inflamed}
\title{Compute T cell-inflamed signature score}
\usage{
compute.Tcell_inflamed(RNA.tpm)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=T cell-inflamed signature score
}
\description{
\code{compute_ayersTcellInfl} computes T cell-inflamed signature score by taking a weighted sum of
the housekeeping normalized values of the T cell-inflamed signature genes
}
