% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.TIDE.R
\name{compute.TIDE}
\alias{compute.TIDE}
\title{Compute TIDE score}
\usage{
compute.TIDE(RNA.tpm, cancertype, output_file_path)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}

\item{cancertype}{string character}
}
\value{
numeric matrix with rows=samples and columns=TIDE score
}
\description{
\code{compute_TIDE} computes TIDE score using tidepy python module. Input data is processed as recommended
by the authors: log2 transformation, quantile normalization and mean centralization (Jiang et al., 2018).
}
