% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.TLS.R
\name{compute.TLS}
\alias{compute.TLS}
\title{Compute tertiary lymphoid structures signature}
\usage{
compute.TLS(RNA.tpm)
}
\arguments{
\item{RNA.tpm}{numeric matrix with rows=genes and columns=samples}
}
\value{
numeric matrix with rows=samples and columns=TLS signature
}
\description{
\code{compute_TLS} computes TLS signature as the geometric-mean of TLS signature genes
(Cabrita et al., 2020).
}
