% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CC_pairs.R
\name{compute_CC_pairs}
\alias{compute_CC_pairs}
\title{compute_CC_pairs}
\usage{
compute_CC_pairs(lrpairs, cancertpe)
}
\arguments{
\item{lrpairs}{Ligand-leceptor pairs weights matrix}

\item{cancertype}{string character}
}
\value{
A list with the following elements:
\describe{
\item{score}{cell-cell interaction scores matrix with rows=samples and columns=cells}
\item{pval}{corresponding p-values for each cell-cell pair, comparing the score with
the null hypothesis}
}
}
\description{
\code{compute_CC_pairs} compute CC pairs from tpm, using the null model for CC interaction computed on TCGA data
}
