% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CCpair_pval.R
\name{compute_CCpair_pval}
\alias{compute_CCpair_pval}
\title{compute_CCpair_pval}
\usage{
compute_CCpair_pval(value, null.hypothesis)
}
\arguments{
\item{value}{CC pair score}

\item{null.hypothesis}{hypothesized distribution for each cell-cell pair}
}
\value{
p-value for each cell-cell pair
}
\description{
\code{compute_CCpair_pval} compute corresponding p-values by comparing the score with the null hypothesis computed for each cell-cell pair.
}
