% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CCpair_score.R
\name{compute_CCpair_score}
\alias{compute_CCpair_score}
\title{compute_CCpair_score}
\usage{
compute_CCpair_score(
  celltype1,
  celltype2,
  intercell.network,
  lrpairs.binary,
  lr.frequency,
  compute.log = T
)
}
\arguments{
\item{celltype1}{string character with first cell type involved in the interaction}

\item{celltype2}{string character with second cell type involved in the interaction}

\item{lrpairs.binary}{binary vector displaying LR pairs with non-zero frequency}

\item{lr.frequency}{numeric vector with LR pairs frequency across the whole TCGA database}

\item{compute.log}{boolean variable in order to take the log of the weighted score}

\item{intracell.network}{matrix with data on cell types interaction network}

\item{cancertype}{string character}
}
\value{
numeric vector with weighted scores
}
\description{
\code{compute_CCpair_score} computes the score for each cell - cell pair for all the patients.
}
