% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pathways_scores.R
\name{compute_pathways_scores}
\alias{compute_pathways_scores}
\title{Compute pathways score}
\usage{
compute_pathways_scores(RNA.counts, remove.genes.ICB_proxies = TRUE)
}
\arguments{
\item{RNA.counts}{numeric matrix of read counts with rows=genes and columns=samples}

\item{remove.genes.ICB_proxies}{boolean variable to reomove all those genes involved in the computation of ICB proxy's of response}
}
\value{
A list with the following elements:
\describe{
\item{scores}{pathway activity matrix with rows=samples and columns=pathwats}
\item{transcripts_kept}{vector with available gene names}
\item{transcripts_left}{vector with missing gene names}
}
}
\description{
\code{compute_pathways_scores} infers pathway activity from raw counts RNAseq data.
}
