% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_TF_activity.R
\name{compute_TF_activity}
\alias{compute_TF_activity}
\title{Compute transription factors activity}
\usage{
compute_TF_activity(RNA.tpm, remove.genes.ICB_proxies = FALSE)
}
\arguments{
\item{RNA.tpm}{numeric matrix of tpm values with rows=genes and columns=samples}

\item{remove.genes.ICB_proxies}{boolean variable to reomove all those genes involved in the computation of ICB proxy's of response}
}
\value{
A list with the following elements:
\describe{
\item{scores}{TF activity matrix with rows=samples and columns=TFs}
\item{transcripts_kept}{vector with available gene names}
\item{transcripts_left}{vector with missing gene names}
}
}
\description{
\code{compute_TF_activity} infers transription factor activity from tpm RNAseq data.
}
