% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_biomarkers.R
\name{explore_biomarkers}
\alias{explore_biomarkers}
\title{Examination of possible relevant biomarkers in cancer dataset}
\usage{
explore_biomarkers(
  pathways,
  immunecells,
  lrpairs,
  cytokinepairs,
  tfs,
  ccpairsgroupedscores,
  ccpairsgroupedpval,
  cancertype,
  real_patient_response,
  output_file_path
)
}
\arguments{
\item{pathways}{numeric matrix with data}

\item{immunecells}{numeric matrix with data}

\item{lrpairs}{numeric matrix with data}

\item{cytokinepairs}{numeric matrix with data}

\item{tfs}{numeric matrix with data}

\item{ccpairsgroupedscores}{numeric matrix with data}

\item{ccpairsgroupedpval}{numeric matrix with data}

\item{cancertype}{string character}

\item{real_patient_response}{vector with two factors (NR,R)}

\item{output_file_path}{string with a file name}
}
\value{
boxplot with features distribution
}
\description{
\code{explore_biomarkers} provides an overview of the mechanistic signatures features distribution in
the dataset (separating responders and non-responders if known) together with information on the
feature contribution to the model building.
}
