% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OE_bulk.R
\name{get_OE_bulk}
\alias{get_OE_bulk}
\title{Calculate overall expression (OE)}
\usage{
get_OE_bulk(r, gene.sign = NULL, num.rounds = 1000, full.flag = FALSE)
}
\arguments{
\item{r}{list}

\item{gene.sign}{string}

\item{num.rounds}{integer}

\item{full.flag}{boolean}
}
\value{
Random score
}
\description{
\code{get_OE_bulk} obtained from literature to calculate Immune resistance program  (Jerby-Arnon et al., 2018)
}
