% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_semi_random_OE.R
\name{get_semi_random_OE}
\alias{get_semi_random_OE}
\title{Calculate random scores}
\usage{
get_semi_random_OE(r, genes.dist.q, b.sign, num.rounds = 1000, full.flag = F)
}
\arguments{
\item{r}{list}

\item{genes.dist.q}{integer}

\item{b.sign}{boolean}

\item{num.rounds}{integer}

\item{full.flag}{boolean}
}
\value{
Random score
}
\description{
\code{get_semi_random_OE} obtained from literature to calculate Immune resistance program  (Jerby-Arnon et al., 2018)
}
