% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsmap.R
\name{ipsmap}
\alias{ipsmap}
\title{Calculate Immunophenoscore value}
\usage{
ipsmap(x)
}
\arguments{
\item{x}{numeric value per sample, needed to check more precisely.}
}
\value{
IPS per sample
}
\description{
\code{ipsmap} obtained from literature to calculate immunophenoscore.  (Charoentong et al., 2017)
}
