% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_task_EN_test.R
\name{multi_task_EN_test}
\alias{multi_task_EN_test}
\title{Multi-task elastic net algorithm implements the model training coefficients and
hyperparameters on the test set.}
\usage{
multi_task_EN_test(x.test, coef.matrix)
}
\arguments{
\item{x.test}{A matrix with data required to perform predictions.}

\item{coef.matrix}{A coefficient matrix obtained during training the model (rows = features and columns = tasks).}
}
\value{
A matrix with predictions for each sample and task.
}
\description{
\code{multi_task_EN_test} implements mutli-task elastic net regression on the test data
}
