% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_immune_response.R
\name{predict_immune_response}
\alias{predict_immune_response}
\title{Immune response prediction}
\usage{
predict_immune_response(
  pathways = NULL,
  immunecells = NULL,
  tfs = NULL,
  lrpairs = NULL,
  ccpairs = NULL,
  cancertype
)
}
\arguments{
\item{pathways}{numeric matrix with data}

\item{immunecells}{numeric matrix with data}

\item{tfs}{numeric matrix with data}

\item{lrpairs}{numeric matrix with data}

\item{ccpairs}{numeric matrix with data}

\item{cancertype}{string character}
}
\value{
Predictions for each model building.
}
\description{
\code{predict_immune_response} predicts immune response using two algorithms: multi-task elastic net
and bayesian efficient multi-kernel algorithm. While BEMKL can exploit information across different input and output datasets,
multi-task elastic net can only do so for response variables. Another advantage of BEMKL is missing data handling, which is not
the case for the other algorithm.
}
\details{
These algorithms use model parameters learned during training on different types of data in
order to compute the immune response.
}
