% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_with_bemkl.R
\name{predict_with_bemkl}
\alias{predict_with_bemkl}
\title{Make predictions using BEMKL}
\usage{
predict_with_bemkl(view_name, view_info, view_data, learned_model)
}
\arguments{
\item{view_name}{input view name}

\item{view_info}{input view information of its composition.}

\item{view_data}{input view data as a list. Each item of the list correponds to a certain view.}

\item{learned_model}{parameters learned during training with cross-validation.}
}
\value{
A matrix with the predictions obtained by applying the model on the view input data
}
\description{
\code{predict_with_bemkl} predicts immune response using bayesian efficient multi-kernel algorithm.
This algorithm employs model parameters learned during training on different types of data in
order to compute the immune response.
}
