% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allData_riaz.R
\docType{data}
\name{Riaz_data}
\alias{Riaz_data}
\title{Transcriptomics, immune cell quantification and patient response data from 49 melanoma patients}
\format{
A list with three dataframes and one character vector.
\describe{
\item{raw_counts_RNAseq}{A dataframe containing gene count data. 19423 genes as rows and 49 samples as columns.}
\item{tpm_RNAseq}{A dataframe containing Transcripts Per Kilobase Million data. 19423 genes as rows and 49 samples as columns.}
\item{cell_fractions}{Using deconvolution approaches on RNA seq data, we estimated cell fractions.}
\item{patient_response}{Response for 49 patient treated with anti-PD-1.}
}
}
\source{
Riaz N, Havel JJ, Makarov V, Desrichard A et al. Tumor and Microenvironment Evolution during Immunotherapy with
Nivolumab. Cell 2017 Nov 2;171(4):934-949.e16. PMID: 29033130
}
\usage{
Riaz_data
}
\description{
Transcriptomics, immune cell quantification and patient response data from 49 melanoma patients
}
\keyword{datasets}
